fix #6503: return api handler errors in the json body
similar to what we do for the extjs formatter, put the error message or
status message in the 'message' property of the return object.

This way client libraries can extract the error without having to parse
the HTTP status reason phrase (which is not possible in all http
libraries, e.g. hyperium's http rust crate).

This should not be a breaking change, since it just adds a (semi) new
field to the return value.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>